refactor(slo-v2): make bare client internal
