+++ clients/slo/internal/client/slo.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package slo
+package client
 
 import (
 	"bytes"
